deleting now requires secret password

+++ controllers/categoryController.js
@@ -2,7 +2,7 @@ import Item from "../models/item.js";
 import Category from "../models/category.js";
 import asyncHandler from "express-async-handler";
 import { body, validationResult } from "express-validator";
-import { render } from "pug";
+import "dotenv/config";
 
 export const categoryList = asyncHandler(async (req, res, next) => {
   const categories = await Category.find()
@@ -128,7 +128,33 @@ export const categoryDelete_get = asyncHandler(async (req, res, next) => {
   });
 });
 
-export const categoryDelete_post = asyncHandler(async (req, res, next) => {
-  await Category.findByIdAndDelete(req.body.categoryid).exec();
-  res.redirect("/inventory/categories");
-});
+export const categoryDelete_post = [
+  body("password", "Incorrect Admin Password").matches(process.env.SECRET_PASSWORD),
+
+  asyncHandler(async (req, res, next) => {
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      const category = await Category.findById(req.params.id)
+        .populate({
+          path: "items",
+          select: "name",
+          options: {
+            collation: { locale: "en", strength: 2 },
+            sort: { name: 1 },
+          },
+        })
+        .exec();
+      res.render(`category_details`, {
+        title: `Delete Category?`,
+        category,
+        toDelete: true,
+        errors: errors.array(),
+      });
+      return;
+    } else {
+      await Category.findByIdAndDelete(req.body.categoryid).exec();
+      res.redirect("/inventory/categories");
+    }
+  }),
+];
